Moving UnitOfWork annotation to com.scottescue.dropwizard.entitymanager
