+++ src/main/java/com/scottescue/dropwizard/entitymanager/UnitOfWork.java
@@ -1,4 +1,4 @@
-package io.dropwizard.hibernate;
+package com.scottescue.dropwizard.entitymanager;
 
 import com.scottescue.dropwizard.entitymanager.EntityManagerBundle;
 import org.hibernate.CacheMode;
