+++ src/main/java/com/scottescue/dropwizard/entitymanager/UnitOfWorkAwareProxyFactory.java
@@ -3,7 +3,6 @@ package com.scottescue.dropwizard.entitymanager;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableMap;
-import io.dropwizard.hibernate.UnitOfWork;
 import javassist.util.proxy.Proxy;
 import javassist.util.proxy.ProxyFactory;
