+++ src/test/java/com/scottescue/dropwizard/entitymanager/UnitOfWorkTest.java
@@ -1,4 +1,4 @@
-package io.dropwizard.hibernate;
+package com.scottescue.dropwizard.entitymanager;
 
 import org.hibernate.CacheMode;
 import org.hibernate.FlushMode;
